Remove obosolete connection upgrade code (now replaced by stdlib proxy)
